[Refactor] 댓글 관련 훅 분리
